commit some changes as stated by amjad bhai

+++ app/src/main/java/edu/aku/hassannaqvi/kmc_screening/ui/form3/SectionEForm3Activity.java
@@ -90,7 +90,7 @@ public class SectionEForm3Activity extends AppCompatActivity {
             @Override
             public void onTextChanged(CharSequence charSequence, int i, int i1, int i2) {
 
-                int no = bi.kf3e04a.getText().toString().isEmpty() ? 0 : Integer.valueOf(bi.kf3e04a.getText().toString());
+                float no = bi.kf3e04a.getText().toString().isEmpty() ? 0f : Float.valueOf(bi.kf3e04a.getText().toString());
 
                 if (no >= 38 || no <= 35.5) {
                     bi.kf3e04b.setVisibility(View.VISIBLE);
